+++ src/jobs/auth-verification-email.ts
@@ -3,7 +3,7 @@ import { env } from "@/env";
 import { sendMail } from "@/server/mailer";
 import { renderAsync } from "@react-email/components";
 import { z } from "zod";
-import { defineJob, defineWorker, defineWorkerConfig } from "./queue";
+import { defineJob, defineWorker, defineWorkerConfig } from "../lib/queue";
 
 const config = defineWorkerConfig({
   name: "email.auth-verify",
