+++ src/jobs/esign-email.ts
@@ -4,7 +4,7 @@ import { db } from "@/server/db";
 import { sendMail } from "@/server/mailer";
 import { renderAsync } from "@react-email/components";
 import { z } from "zod";
-import { defineJob, defineWorker, defineWorkerConfig } from "./queue";
+import { defineJob, defineWorker, defineWorkerConfig } from "../lib/queue";
 
 const Schema = z.object({
   documentName: z.string().optional(),
