+++ src/jobs/member-inivite-email.ts
@@ -4,7 +4,7 @@ import { constants } from "@/lib/constants";
 import { sendMail } from "@/server/mailer";
 import { renderAsync } from "@react-email/components";
 import { z } from "zod";
-import { defineJob, defineWorker, defineWorkerConfig } from "./queue";
+import { defineJob, defineWorker, defineWorkerConfig } from "../lib/queue";
 
 const config = defineWorkerConfig({
   name: "email.member-invite",
